requirements.txt
appdirs
astroid
attrs
black
cfgv
click
distlib
filelock
#freeze
identify
iniconfig
isort
lazy-object-proxy
mccabe
mypy-extensions
nodeenv
packaging
pathspec
pluggy
pre-commit
py
pylint
pyparsing
pytest
PyYAML
regex
six
virtualenv
wrapt
pandas
numpy
matplotlib
seaborn
scipy
sqlalchemy
psycopg2
pg8000
python-dotenv
requests
